Add files via upload
In this simplified version, the computer selects a random number between 1 and 10. The player has 3 attempts to guess the number. After each guess, the program provides feedback on whether the guess is too high or too low. If the player doesn't guess the number within 3 attempts, the program reveals the secret number.

This version is very straightforward and easy to understand. You can run the Python code in an environment that supports user input to play the game. Have fun guessing the number!